bean的字段标注是否可空

+++ sandbox_town_backend/src/main/java/com/shuidun/sandbox_town_backend/mixin/Constants.java
@@ -35,30 +35,30 @@ public class Constants {
     public static final int ITEM_BAR_SIZE = 6;
 
     /** 白天时长5分钟（300000ms） */
-    public static final int DAY_DURATION = 300000;
+    public static final long DAY_DURATION = 300000;
 
     /** 黄昏时长1分钟（60000ms） */
-    public static final int DUSK_DURATION = 60000;
+    public static final long DUSK_DURATION = 60000;
 
     /** 夜晚时长3分钟（180000ms） */
-    public static final int NIGHT_DURATION = 180000;
+    public static final long NIGHT_DURATION = 180000;
 
     /** 黎明时长1分钟（60000ms） */
-    public static final int DAWN_DURATION = 60000;
+    public static final long DAWN_DURATION = 60000;
 
     /** 一天总时长 */
-    public static final int DAY_TOTAL_DURATION = DAY_DURATION + DUSK_DURATION + NIGHT_DURATION + DAWN_DURATION;
+    public static final long DAY_TOTAL_DURATION = DAY_DURATION + DUSK_DURATION + NIGHT_DURATION + DAWN_DURATION;
 
     /** 白天开始时间 */
-    public static final int DAY_START = 0;
+    public static final long DAY_START = 0;
 
     /** 黄昏开始时间 */
-    public static final int DUSK_START = DAY_START + DAY_DURATION;
+    public static final long DUSK_START = DAY_START + DAY_DURATION;
 
     /** 夜晚开始时间 */
-    public static final int NIGHT_START = DUSK_START + DUSK_DURATION;
+    public static final long NIGHT_START = DUSK_START + DUSK_DURATION;
 
     /** 黎明开始时间 */
-    public static final int DAWN_START = NIGHT_START + NIGHT_DURATION;
+    public static final long DAWN_START = NIGHT_START + NIGHT_DURATION;
 
 }
